+++ tests/test_formats.py
@@ -15,8 +15,8 @@ def test_get_cwl_file_format_tuple():
         ns, fmt = res
         assert isinstance(ns, dict) and len(ns) == 1
         assert any(f in ns for f in FORMAT_NAMESPACES)
-        assert ns.values()[0].startswith("http")
-        ns_name = ns.keys()[0]
+        assert list(ns.values())[0].startswith("http")
+        ns_name = list(ns.keys())[0]
         assert fmt.startswith("{}:".format(ns_name))
         tested.remove(ns_name)
     assert len(tested) == 0, "test did not evaluate every namespace variation"
@@ -27,7 +27,7 @@ def test_get_cwl_file_format_reference():
     tests = [(IANA_NAMESPACE_DEFINITION, CONTENT_TYPE_APP_JSON), (EDAM_NAMESPACE_DEFINITION, CONTENT_TYPE_APP_NETCDF)]
     for ns, mime_type in tests:
         res = get_cwl_file_format(mime_type, make_reference=True)
-        ns_name, ns_url = ns.items()[0]
+        ns_name, ns_url = list(ns.items())[0]
         assert isinstance(res, six.string_types)
         assert res.startswith(ns_url)
         tested.remove(ns_name)
@@ -58,16 +58,16 @@ def test_clean_mime_type_format_iana():
     iana_fmt = "{}:{}".format(IANA_NAMESPACE, CONTENT_TYPE_APP_JSON)  # "iana:mime_type"
     res_type = clean_mime_type_format(iana_fmt)
     assert res_type == CONTENT_TYPE_APP_JSON
-    iana_fmt = os.path.join(IANA_NAMESPACE_DEFINITION.values()[0], CONTENT_TYPE_APP_JSON)  # "iana-url/mime_type"
+    iana_fmt = os.path.join(list(IANA_NAMESPACE_DEFINITION.values())[0], CONTENT_TYPE_APP_JSON)  # "iana-url/mime_type"
     res_type = clean_mime_type_format(iana_fmt)
     assert res_type == CONTENT_TYPE_APP_JSON  # application/json
 
 
 def test_clean_mime_type_format_edam():
-    mime_type, fmt = EDAM_MAPPING.items()[0]
+    mime_type, fmt = list(EDAM_MAPPING.items())[0]
     edam_fmt = "{}:{}".format(EDAM_NAMESPACE, fmt)  # "edam:format_####"
     res_type = clean_mime_type_format(edam_fmt)
     assert res_type == mime_type
-    edam_fmt = os.path.join(EDAM_NAMESPACE_DEFINITION.values()[0], fmt)  # "edam-url/format_####"
+    edam_fmt = os.path.join(list(EDAM_NAMESPACE_DEFINITION.values())[0], fmt)  # "edam-url/format_####"
     res_type = clean_mime_type_format(edam_fmt)
     assert res_type == mime_type  # application/x-type